+++ src/main/java/com/vaadin/flow/component/upload/NoInputStreamEvent.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2000-2017 Vaadin Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package com.vaadin.flow.component.upload;
+
+/**
+ * FailedEvent that indicates that an input stream could not be obtained.
+ */
+public class NoInputStreamEvent extends FailedEvent {
+
+    /**
+     *
+     * @param source
+     *            the source of the file.
+     * @param filename
+     *            the received file name.
+     * @param MIMEType
+     *            the MIME type of the received file.
+     * @param length
+     *            the length of the received file.
+     */
+    public NoInputStreamEvent(Upload source, String filename, String MIMEType,
+            long length) {
+        super(source, filename, MIMEType, length);
+    }
+
+}
